Sample the time of the oldest message in a queue (#4)
By deducting this gauge metric value from the current wallclock time, a
Prometheus server may be configured to alert on queues that contain one
or more unusually old messages.  Decrepit messages may be indicative of
a deeper problem that warrants operator intervention.

Postfix uses i-node mtime for its own purposes.  (mtime on files in the
deferred queue, for example, is set to some moment in the future.  These
mtimes are used to signify the moment when a message becomes eligible
for retry.)  AFAICS, Postfix does not overload i-node ctime semantics,
so it should be reasonable to sample that metadata for the purposes of
queue monitoring.

Of course, the ctime of a file in a Postfix queue directory is not even
remotely representative of true message age.  Should Postfix run into a
prolonged spot of bother while attempting to deliver a message to some
destination, a message may ricochet between the active and deferred
queues many times over the course of several days.  File ctime would be
reset on each ricochet.  We trust that operators will pay careful
consideration to Postfix's queuing behaviours when deciding on an
alerting threshold that is fit for their site configuration.